async_lock

Module once_cell

Structs

OnceCell

A memory location that can be written to at most once.

Enums

State 🔒

The current state of the OnceCell.

Functions

now_or_never 🔒

Either return the result of a future now, or panic.