async_lock::rwlock

Module raw

Raw, unsafe reader-writer locking implementation, doesn’t depend on the data protected by the lock. RwLock is implemented in terms of this.

Splitting the implementation this way allows instantiating the locking code only once, and also lets us make RwLockReadGuard covariant in T.

Structs

RawRead 🔒

The future returned by RawRwLock::read.

RawRwLock 🔒

A “raw” RwLock that doesn’t hold any data.

RawUpgradableRead 🔒

The future returned by RawRwLock::upgradable_read.

RawUpgrade 🔒

The future returned by RawRwLock::upgrade.

RawWrite 🔒

The future returned by RawRwLock::write.

Enums

WriteState 🔒

Constants

ONE_READER 🔒

WRITER_BIT 🔒