async_lock

Struct Mutex

pub struct Mutex<T: ?Sized> { /* private fields */ }

An async mutex.

The locking mechanism uses eventual fairness to ensure locking will be fair on average without sacrificing performance. This is done by forcing a fair lock whenever a lock operation is starved for longer than 0.5 milliseconds.

Examples

use async_lock::Mutex;

let m = Mutex::new(1);

let mut guard = m.lock().await;
*guard = 2;

assert!(m.try_lock().is_none());
drop(guard);
assert_eq!(*m.try_lock().unwrap(), 2);

Implementations

impl<T> Mutex<T>

pub const fn new(data: T) -> Mutex<T>

Creates a new async mutex.

Examples

use async_lock::Mutex;

let mutex = Mutex::new(0);

pub fn into_inner(self) -> T

Consumes the mutex, returning the underlying data.

Examples

use async_lock::Mutex;

let mutex = Mutex::new(10);
assert_eq!(mutex.into_inner(), 10);

impl<T: ?Sized> Mutex<T>

pub fn lock(&self) -> Lock<'_, T>

Acquires the mutex.

Returns a guard that releases the mutex when dropped.

Examples

use async_lock::Mutex;

let mutex = Mutex::new(10);
let guard = mutex.lock().await;
assert_eq!(*guard, 10);

pub fn try_lock(&self) -> Option<MutexGuard<'_, T>>

Attempts to acquire the mutex.

If the mutex could not be acquired at this time, then None is returned. Otherwise, a guard is returned that releases the mutex when dropped.

Examples

use async_lock::Mutex;

let mutex = Mutex::new(10);
if let Some(guard) = mutex.try_lock() {
    assert_eq!(*guard, 10);
}

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying data.

Since this call borrows the mutex mutably, no actual locking takes place – the mutable borrow statically guarantees the mutex is not already acquired.

Examples

use async_lock::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut() = 10;
assert_eq!(*mutex.lock().await, 10);

impl<T: ?Sized> Mutex<T>

pub fn lock_arc(self: &Arc<Self>) -> LockArc<T>

Acquires the mutex and clones a reference to it.

Returns an owned guard that releases the mutex when dropped.

Examples

use async_lock::Mutex;
use std::sync::Arc;

let mutex = Arc::new(Mutex::new(10));
let guard = mutex.lock_arc().await;
assert_eq!(*guard, 10);

pub fn try_lock_arc(self: &Arc<Self>) -> Option<MutexGuardArc<T>>

Attempts to acquire the mutex and clone a reference to it.

If the mutex could not be acquired at this time, then None is returned. Otherwise, an owned guard is returned that releases the mutex when dropped.

Examples

use async_lock::Mutex;
use std::sync::Arc;

let mutex = Arc::new(Mutex::new(10));
if let Some(guard) = mutex.try_lock() {
    assert_eq!(*guard, 10);
}