async_lock

Module rwlock

Modules

futures 🔒

raw 🔒

Raw, unsafe reader-writer locking implementation, doesn’t depend on the data protected by the lock. RwLock is implemented in terms of this.

Structs

RwLock

An async reader-writer lock.

RwLockReadGuard

A guard that releases the read lock when dropped.

RwLockReadGuardArc

An owned, reference-counting guard that releases the read lock when dropped.

RwLockUpgradableReadGuard

A guard that releases the upgradable read lock when dropped.

RwLockUpgradableReadGuardArc

An owned, reference-counting guard that releases the upgradable read lock when dropped.

RwLockWriteGuard

A guard that releases the write lock when dropped.

RwLockWriteGuardArc

An owned, reference-counted guard that releases the write lock when dropped.