async_lock::rwlock::raw

Struct RawRwLock

pub(super) struct RawRwLock {
    mutex: Mutex<()>,
    no_readers: Event,
    no_writer: Event,
    state: AtomicUsize,
}

A “raw” RwLock that doesn’t hold any data.

Fields
mutex: Mutex<()>

Acquired by the writer.

no_readers: Event

Event triggered when the last reader is dropped.

no_writer: Event

Event triggered when the writer is dropped.

state: AtomicUsize

Current state of the lock.

The least significant bit (WRITER_BIT) is set to 1 when a writer is holding the lock or trying to acquire it.

The upper bits contain the number of currently active readers. Each active reader increments the state by ONE_READER.

Implementations

impl RawRwLock

pub(super) const fn new() -> Self

pub(super) fn try_read(&self) -> bool

Returns true iff a read lock was successfully acquired.

pub(super) fn read(&self) -> RawRead<'_>

pub(super) fn try_upgradable_read(&self) -> bool

Returns true iff an upgradable read lock was successfully acquired.

pub(super) fn upgradable_read(&self) -> RawUpgradableRead<'_>

pub(super) fn try_write(&self) -> bool

Returns true iff a write lock was successfully acquired.

pub(super) fn write(&self) -> RawWrite<'_>

pub(super) unsafe fn try_upgrade(&self) -> bool

Returns true iff a the upgradable read lock was successfully upgraded to a write lock.

Safety

Caller must hold an upgradable read lock. This will attempt to upgrade it to a write lock.

pub(super) unsafe fn upgrade(&self) -> RawUpgrade<'_>

Safety

Caller must hold an upgradable read lock. This will upgrade it to a write lock.

pub(super) unsafe fn downgrade_upgradable_read(&self)

Safety

Caller must hold an upgradable read lock. This will downgrade it to a standard read lock.

pub(super) unsafe fn downgrade_write(&self)

Safety

Caller must hold a write lock. This will downgrade it to a read lock.

pub(super) unsafe fn downgrade_to_upgradable(&self)

Safety

Caller must hold a write lock. This will downgrade it to an upgradable read lock.

pub(super) unsafe fn read_unlock(&self)

Safety

Caller must hold a read lock . This will unlock that lock.

pub(super) unsafe fn upgradable_read_unlock(&self)

Safety

Caller must hold an upgradable read lock. This will unlock that lock.

pub(super) unsafe fn write_unlock(&self)

Safety

Caller must hold a write lock. This will unlock that lock.